+++ miki7/src/main/java/com/example/miki7/movie/model/SearchDto.java
@@ -1,0 +1,4 @@
+package com.example.miki7.movie.model;
+
+public class SearchDto {
+}
