+++ miki7/src/main/resources/templates/result.html
@@ -1,0 +1,98 @@
+<!DOCTYPE html>
+<html xmlns:th="http://www.thymeleaf.org"
+      xmlns:sec="http://www.thymeleaf.org/extras/spring-security">
+<head>
+    <meta charset="utf-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1, shrink-to-fit=no">
+    <meta http-equiv="X-UA-Compatible" content="ie=edge">
+    <title>Home - Pintereso Bootstrap Template</title>
+    <script type="text/javascript">
+        (function () {
+            var css = document.createElement('link');
+            css.href = 'https://stackpath.bootstrapcdn.com/font-awesome/4.7.0/css/font-awesome.min.css';
+            css.rel = 'stylesheet';
+            css.type = 'text/css';
+            document.getElementsByTagName('head')[0].appendChild(css);
+        })();
+    </script>
+    <link rel="stylesheet" th:href="@{/css/app.css}">
+    <link rel="stylesheet" th:href="@{/css/theme.css}">
+</head>
+<body>
+
+<!-- Topbar 포함 -->
+<div th:insert="~{/fragments/topbar :: topbar}"></div>
+
+<main role="main">
+
+    <section class="mt-4 mb-5">
+        <div class="container mb-4">
+            <h1 class="font-weight-bold title">Explore</h1>
+            <div class="row">
+                <nav class="navbar navbar-expand-lg navbar-light bg-white pl-2 pr-2">
+                    <button class="navbar-light navbar-toggler" type="button" data-toggle="collapse"
+                            data-target="#navbarsExplore" aria-controls="navbarsDefault" aria-expanded="false"
+                            aria-label="Toggle navigation">
+                        <span class="navbar-toggler-icon"></span>
+                    </button>
+                    <div class="collapse navbar-collapse" id="navbarsExplore">
+                        <ul class="navbar-nav">
+                            <li class="nav-item">
+                                <a class="nav-link" href="#">Lifestyle</a>
+                            </li>
+                            <li class="nav-item">
+                                <a class="nav-link" href="#">Food</a>
+                            </li>
+                            <li class="nav-item">
+                                <a class="nav-link" href="#">Home</a>
+                            </li>
+                            <li class="nav-item">
+                                <a class="nav-link" href="#">Travel</a>
+                            </li>
+                            <li class="nav-item dropdown">
+                                <a class="nav-link dropdown-toggle" href="http://example.com" id="dropdown01"
+                                   data-toggle="dropdown" aria-haspopup="true" aria-expanded="false">More</a>
+                                <div class="dropdown-menu shadow-lg" aria-labelledby="dropdown01">
+                                    <a class="dropdown-item" href="#">Astronomy</a>
+                                    <a class="dropdown-item" href="#">Nature</a>
+                                    <a class="dropdown-item" href="#">Cooking</a>
+                                    <a class="dropdown-item" href="#">Fashion</a>
+                                    <a class="dropdown-item" href="#">Wellness</a>
+                                    <a class="dropdown-item" href="#">Dieting</a>
+                                </div>
+                            </li>
+                        </ul>
+                    </div>
+                </nav>
+            </div>
+        </div>
+        <div class="container-fluid">
+            <div class="row">
+                <div class="card-columns">
+                    <div class="card card-pin">
+                        <img class="card-img" th:src="@{/img/card1.jpg}" alt="Card image">
+                        <div class="overlay">
+                            <h2 class="card-title title">Cool Title</h2>
+                            <div class="more">
+                                <a th:href="@{/post}">
+                                    <i class="fa fa-arrow-circle-o-right" aria-hidden="true"></i> More
+                                </a>
+                            </div>
+                        </div>
+                    </div>
+                    <!-- 다른 카드들 -->
+                </div>
+            </div>
+        </div>
+    </section>
+
+</main>
+
+<script th:src="@{/js/app.js}"></script>
+<script th:src="@{/js/theme.js}"></script>
+
+<!-- Footer 포함 -->
+<div th:insert="~{/fragments/footer :: footer}"></div>
+</body>
+
+</html>
